[FEAT] 23/09/19 BLUEWAVE 마이페이지 비공개 필터링 수정, 에러 메세지 수정 중. 그외
1. 구독 프로필 이미지 링크 수정
2. 구독 TBL PK 수정
3. 타 회원 마이페이지 조회 시 비공개 챌린지 필터링 추가
	- 비공개 개인 챌린지는 본인만 조회 가능
4. 챌린지 생성 시 챌린지 페이지로 이동
5. 챌린지 CRUD 시 세션 아이디 체크 중
6. 에러 메세지 통일하게 수정 중..

+++ src/main/java/com/kh/bluewave/challenge/controller/ChallengeController.java
@@ -61,43 +61,49 @@ public class ChallengeController {
 				chalEndDate = new java.sql.Date(Date.valueOf(endDate).getTime());
 			}
 			
-			// 로그인 여부 확인 (추후 수정)
-			challenge.setChalUserId("testuser01"); // 임시 사용자 아이디
-			
-			// 챌린지 명 중복 여부 확인
-			Challenge cOne = cService.selectOneByTitle(challenge.getChalTitle());
-			if(cOne == null) {
-				// 파일 첨부
-				if(!uploadFile.getOriginalFilename().equals("")) {
-					Map<String , Object> cMap = this.saveFile(request, uploadFile);
-					String fileName = (String)cMap.get("fileName"); 
-					String fileRename = (String)cMap.get("fileRename");
-					String savePath = (String)cMap.get("filePath");
-					long fileLength = (long)cMap.get("fileLength");
+			// 로그인 여부 확인
+			if(challenge.getChalUserId() != null) {
+				// 챌린지 명 중복 여부 확인
+				Challenge cOne = cService.selectOneByTitle(challenge.getChalTitle());
+				if(cOne == null) {
+					// 파일 첨부
+					if(!uploadFile.getOriginalFilename().equals("")) {
+						Map<String , Object> cMap = this.saveFile(request, uploadFile);
+						String fileName = (String)cMap.get("fileName"); 
+						String fileRename = (String)cMap.get("fileRename");
+						String savePath = (String)cMap.get("filePath");
+						long fileLength = (long)cMap.get("fileLength");
+						
+						challenge.setChalFileName(fileName);
+						challenge.setChalFileRename(fileRename);
+						challenge.setChalFilePath(savePath);
+						challenge.setChalFileLength(fileLength);
+					} else {
+						mv.addObject("msg", "파일 등록 실패").addObject("url", "/challenge/page.do");
+						mv.setViewName("common/serviceFailed");
+					}
+					challenge.setChalStartDate(chalStartDate);
+					challenge.setChalEndDate(chalEndDate);
 					
-					challenge.setChalFileName(fileName);
-					challenge.setChalFileRename(fileRename);
-					challenge.setChalFilePath(savePath);
-					challenge.setChalFileLength(fileLength);
-				}
-				challenge.setChalStartDate(chalStartDate);
-				challenge.setChalEndDate(chalEndDate);
-				
-				int result = cService.insertChal(challenge);
-				if(result > 0) {
-					mv.setViewName("/user/myPage");				
+					int result = cService.insertChal(challenge);
+					if(result > 0) {
+						mv.setViewName("/challenge/page.do");				
+					} else {
+						mv.addObject("msg", "챌린지 생성 오류").addObject("url", "/challenge/page.do");
+						mv.setViewName("common/serviceFailed");
+					}
 				} else {
-					mv.addObject("msg", "챌린지 생성 오류");
-					mv.setViewName("common/errorMessage");
+					mv.addObject("msg", "챌린지 명 중복").addObject("url", "/challenge/page.do");
+					mv.setViewName("common/serviceFailed");
 				}
 			} else {
-				mv.addObject("msg", "챌린지 명 중복");
-				mv.setViewName("common/errorMessage");
+				mv.addObject("msg", "로그인 필요").addObject("url", "/user/login.do");
+				mv.setViewName("common/serviceFailed");
 			}
 		} catch (Exception e) {
 			e.printStackTrace();
-			mv.addObject("msg", e.getMessage());
-			mv.setViewName("common/errorMessage");
+			mv.addObject("msg", "챌린지 생성 실패").addObject("url", "/challenge/page.do");
+			mv.setViewName("common/serviceFailed");
 		}
 		return mv;
 	}
@@ -112,20 +118,23 @@ public class ChallengeController {
 	
 	// 챌린지 수정 페이지
 	@RequestMapping(value="/challenge/update.do", method=RequestMethod.GET)
-	public ModelAndView showUpdateChalForm(int chalNo, ModelAndView mv) {
+	public ModelAndView showUpdateChalForm(int chalNo
+										 , String chalUserId
+										 , String userId
+										 , ModelAndView mv) {
 		try {
 			Challenge cOne = cService.selectOneByNo(chalNo);
 			if(cOne != null) {
 				mv.addObject("chal", cOne);
 				mv.setViewName("/challenge/updateChal");
 			} else {
-				mv.addObject("msg", "수정할 챌린지 데이터 불러오기");
-				mv.setViewName("common/errorMessage");
+				mv.addObject("msg", "수정할 챌린지 데이터 불러오기").addObject("url", "/user/myPage.do?userId=" + userId);
+				mv.setViewName("common/serviceFailed");
 			}
 		} catch (Exception e) {
 			e.printStackTrace();
-			mv.addObject("msg", e.getMessage());
-			mv.setViewName("common/errorMessage");
+			mv.addObject("msg", "수정 페이지 로딩 실패").addObject("url", "/challenge/info.do?chalNo=" + chalNo);
+			mv.setViewName("common/serviceFailed");
 		}
 		return mv;
 	}
@@ -133,41 +142,56 @@ public class ChallengeController {
 	// 챌린지 수정 기능
 	@RequestMapping(value="/challenge/update.do", method=RequestMethod.POST)
 	public ModelAndView updateChal(@ModelAttribute Challenge challenge
+								   , String userId
 								   , @RequestParam(value="uploadFile", required=false) MultipartFile uploadFile
 								   , HttpServletRequest request
 								   , RedirectAttributes redirect
 								   , ModelAndView mv) {
 		try {
-			if(!uploadFile.getOriginalFilename().equals("")) {
-				String fileRename = challenge.getChalFileRename();
-				if(fileRename != null) {
-					this.deleteFile(request, fileRename);
+			// 로그인 여부 및 본인 일치 여부 확인
+			if(userId != null) {
+				if(challenge.getChalUserId() == userId) {
+					if(!uploadFile.getOriginalFilename().equals("")) {
+						String fileRename = challenge.getChalFileRename();
+						if(fileRename != null) {
+							this.deleteFile(request, fileRename);
+						}
+						
+						Map<String , Object> cMap = this.saveFile(request, uploadFile);
+						String chalFileName = (String)cMap.get("fileName"); 
+						String chalFileRename = (String)cMap.get("fileRename");
+						String chalFilePath = (String)cMap.get("filePath");
+						long chalFileLength = (long)cMap.get("fileLength");
+						
+						challenge.setChalFileName(chalFileName);
+						challenge.setChalFileRename(chalFileRename);
+						challenge.setChalFilePath(chalFilePath);
+						challenge.setChalFileLength(chalFileLength);
+					} else {
+						mv.addObject("msg", "파일 수정 실패").addObject("url", "/user/myPage.do?userId=" + userId);
+						mv.setViewName("common/serviceFailed");
+					}
+					
+					int result = cService.updateChal(challenge);
+					
+					if(result > 0) {
+						mv.setViewName("/user/myPage");		
+					} else {
+						mv.addObject("msg", "챌린지 수정 실패").addObject("url", "/user/myPage.do?userId=" + userId);
+						mv.setViewName("common/serviceFailed");
+					}
+				} else {
+					mv.addObject("msg", "작성자만 챌린지 수정 가능").addObject("url", "/user/login.do");
+					mv.setViewName("common/serviceFailed");
 				}
-				
-				Map<String , Object> cMap = this.saveFile(request, uploadFile);
-				String chalFileName = (String)cMap.get("fileName"); 
-				String chalFileRename = (String)cMap.get("fileRename");
-				String chalFilePath = (String)cMap.get("filePath");
-				long chalFileLength = (long)cMap.get("fileLength");
-				
-				challenge.setChalFileName(chalFileName);
-				challenge.setChalFileRename(chalFileRename);
-				challenge.setChalFilePath(chalFilePath);
-				challenge.setChalFileLength(chalFileLength);
-			}
-			
-			int result = cService.updateChal(challenge);
-			
-			if(result > 0) {
-				mv.setViewName("/user/myPage");		
 			} else {
-				mv.addObject("msg", "챌린지 수정");
-				mv.setViewName("common/errorMessage");
+				mv.addObject("msg", "로그인 필요").addObject("url", "/user/login.do");
+				mv.setViewName("common/serviceFailed");
 			}
 		} catch (Exception e) {
 			e.printStackTrace();
-			mv.addObject("msg", e.getMessage());
-			mv.setViewName("common/errorMessage");
+			mv.addObject("msg", "수정 페이지 로딩 실패").addObject("url", "/user/myPage.do?userId=" + userId);
+			mv.setViewName("common/serviceFailed");
 		}
 		return mv;
 	}
@@ -175,22 +199,35 @@ public class ChallengeController {
 	// 챌린지 삭제 기능
 	@RequestMapping(value="/challenge/delete.do", method=RequestMethod.GET)
 	public ModelAndView deleteChal(int chalNo
+								, String chalUserId
+								, String userId
 								, HttpServletRequest request
 								, ModelAndView mv) {
 		try {
-			Challenge challenge = cService.selectOneByNo(chalNo);
-			int result = cService.deleteChal(chalNo);
-			if(result > 0) {
-				this.deleteFile(request, challenge.getChalFileRename());
-				mv.setViewName("/user/myPage");
+			// 로그인 여부 및 본인 일치 여부 확인
+			if(userId != null) {
+				if(chalUserId == userId) {
+					Challenge challenge = cService.selectOneByNo(chalNo);
+					int result = cService.deleteChal(chalNo);
+					if(result > 0) {
+						this.deleteFile(request, challenge.getChalFileRename());
+						mv.setViewName("/user/myPage");
+					} else {
+						mv.addObject("msg", "챌린지 삭제 실패").addObject("url", "/user/myPage.do?userId=" + userId);
+						mv.setViewName("common/serviceFailed");
+					}
+				} else {
+					mv.addObject("msg", "작성자만 챌린지 삭제 가능").addObject("url", "/user/login.do");
+					mv.setViewName("common/serviceFailed");
+				}
 			} else {
-				mv.addObject("msg", "챌린지 삭제");
-				mv.setViewName("common/errorMessage");
+				mv.addObject("msg", "로그인 필요").addObject("url", "/user/login.do");
+				mv.setViewName("common/serviceFailed");
 			}
 		} catch (Exception e) {
 			e.printStackTrace();
-			mv.addObject("msg", e.getMessage());
-			mv.setViewName("common/errorMessage");
+			mv.addObject("msg", "챌린지 삭제 실패").addObject("url", "/user/myPage.do?userId=" + userId);
+			mv.setViewName("common/serviceFailed");
 		}
 		return mv;
 	}
@@ -271,7 +308,7 @@ public class ChallengeController {
 		} catch (Exception e) {
 			e.printStackTrace();
 			mv.addObject("msg", e.getMessage());
-			mv.setViewName("common/errorMessage");
+			mv.setViewName("common/serviceFailed");
 		}
 		mv.setViewName("challenge/challengePage");
 		return mv;
